mainにloggerを統合
